requirements.txt
# ============================================================
# Blockchain-Enabled Adaptive Weighted FedAvg
# requirements.txt  –  core runtime dependencies only
# ============================================================
#
# Install with:
#   pip install -r requirements.txt
#
# Python: 3.9 – 3.11 recommended
# ============================================================

# ── Federated Learning ──────────────────────────────────────
flwr[simulation]>=1.5.0,<2.0.0   # [simulation] pulls in ray for fl.simulation.start_simulation
stable-baselines3>=2.0.0,<3.0.0

# ── Reinforcement Learning / Environment ────────────────────
gymnasium>=0.29.0,<1.0.0

# ── Deep Learning ───────────────────────────────────────────
# CUDA build: replace with:
#   pip install torch>=2.0.0,<2.4.0 --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0,<2.4.0

# ── Numerics ────────────────────────────────────────────────
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# ── Blockchain ──────────────────────────────────────────────
# web3 6.x ships strict upper bounds for eth-account (<0.10.0)
# and hexbytes (<0.4.0). These bounds MUST be reflected here or
# pip will pick eth-account 0.10+ / hexbytes 0.4+ and fail.
web3>=6.0.0,<7.0.0
eth-account>=0.9.0,<0.10.0
eth-utils>=2.2.0,<3.0.0
hexbytes>=0.3.0,<0.4.0

# ── IPFS ────────────────────────────────────────────────────
# ipfshttpclient 0.8.0a2 is the last release and requires
# multiaddr < 2.0.0. Pin both to prevent silent upgrades.
ipfshttpclient==0.8.0a2
multiaddr>=0.0.9,<2.0.0

# ── Cryptography / Privacy ──────────────────────────────────
# Do not hard-pin the patch: let pip negotiate with web3's own
# cryptography lower bound (>=35.0.0).  Cap at <43 to avoid
# API breaks in the pyOpenSSL / cryptography chain.
cryptography>=35.0.0,<43.0.0
pycryptodome>=3.18.0

# ── System / Resource Monitoring ────────────────────────────
psutil>=5.9.0
GPUtil>=1.4.0
memory-profiler>=0.61.0

# ── Data / Utilities ────────────────────────────────────────
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0.1

# ── Visualisation ───────────────────────────────────────────
matplotlib>=3.7.0,<3.10.0
pyparsing>=2.4.7,<3.0.0         # matplotlib internals use old pyparsing API; 3.x breaks it
seaborn>=0.12.0

# ── Smart contract tooling ──────────────────────────────────
py-solc-x>=1.1.1
